chore: reorganizing files
- move service files to utils
