+++ app/routes/_auth+/logout.tsx
@@ -1,7 +1,7 @@
 import type { ActionFunctionArgs } from "@remix-run/node";
 import { redirect } from "@remix-run/node";
 
-import { logout } from "~/session.server";
+import { logout } from "~/utils/session.server";
 
 export const action = async ({ request }: ActionFunctionArgs) =>
   logout(request);
