+++ app/routes/_marketing+/_index.tsx
@@ -1,7 +1,7 @@
 import type { MetaFunction } from "@remix-run/node";
 import { Link } from "@remix-run/react";
 
-import { useOptionalUser } from "~/utils";
+import { useOptionalUser } from "~/utils/utils";
 
 export const meta: MetaFunction = () => [{ title: "Remix Notes" }];
