+++ app/routes/healthcheck.tsx
@@ -1,6 +1,6 @@
 import type { LoaderFunctionArgs } from "@remix-run/node";
 
-import { prisma } from "~/db.server";
+import { prisma } from "~/utils/db.server";
 
 export const loader = async ({ request }: LoaderFunctionArgs) => {
   const host =
