+++ app/routes/notes+/$noteId.tsx
@@ -8,8 +8,8 @@ import {
 } from "@remix-run/react";
 import invariant from "tiny-invariant";
 
-import { deleteNote, getNote } from "~/models/note.server";
-import { requireUserId } from "~/session.server";
+import { deleteNote, getNote } from "~/utils/note.server";
+import { requireUserId } from "~/utils/session.server";
 
 export const loader = async ({ params, request }: LoaderFunctionArgs) => {
   const userId = await requireUserId(request);
