+++ app/routes/notes+/new.tsx
@@ -3,8 +3,8 @@ import { json, redirect } from "@remix-run/node";
 import { Form, useActionData } from "@remix-run/react";
 import { useEffect, useRef } from "react";
 
-import { createNote } from "~/models/note.server";
-import { requireUserId } from "~/session.server";
+import { createNote } from "~/utils/note.server";
+import { requireUserId } from "~/utils/session.server";
 
 export const action = async ({ request }: ActionFunctionArgs) => {
   const userId = await requireUserId(request);
