+++ app/utils/session.server.ts
@@ -1,8 +1,8 @@
 import { createCookieSessionStorage, redirect } from "@remix-run/node";
 import invariant from "tiny-invariant";
 
-import type { User } from "~/models/user.server";
-import { getUserById } from "~/models/user.server";
+import type { User } from "~/utils/user.server";
+import { getUserById } from "~/utils/user.server";
 
 invariant(process.env.SESSION_SECRET, "SESSION_SECRET must be set");
