+++ app/utils/singleton.server.ts
