+++ app/utils/utils.ts
@@ -1,7 +1,7 @@
 import { useMatches } from "@remix-run/react";
 import { useMemo } from "react";
 
-import type { User } from "~/models/user.server";
+import type { User } from "~/utils/user.server";
 
 const DEFAULT_REDIRECT = "/";
